+++ Samples~/CSV Example/Editor/ExampleCSVStreamingAssetsImporter.cs
@@ -1,0 +1,29 @@
+namespace rlmg.Tools.ContentLoading.Examples.Editor
+{
+    using UnityEditor;
+    using UnityEditor.SceneManagement;
+    using UnityEngine.SceneManagement;
+
+    [InitializeOnLoad]
+    public static class ExampleCSVStreamingAssetsImporter
+    {
+        static ExampleCSVStreamingAssetsImporter()
+        {
+            EditorSceneManager.sceneOpened += OnSceneOpened;
+        }
+
+        private static void OnSceneOpened(Scene scene, OpenSceneMode mode)
+        {
+
+            if (scene.name.Contains("ExampleCSVLoader"))
+                ImportStreamingAssets();
+        }
+        private static void ImportStreamingAssets()
+        {
+            SampleStreamingAssetsImporterUtils.CopySampleFilesToStreamingAssets(
+                importerScriptName: nameof(ExampleCSVStreamingAssetsImporter),
+                destStreamingAssetsSubFolderName: "RLMG Content Loading Samples/CSV");
+        }
+    }
+
+}