+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.1]
+
+### Changed
+
+- all Samples' loaders' local file paths
+- ExampleSequentialLoader no longer references UI displays
+
+### Added
+
+- HiddenStreamingAssets~ folders and supporting Editor scripts
+- ExampleSequentialLoaderListener
+
+### Removed
+
+- streamingassets.unitypackage
+
 ## [0.3.0]
 
 ### Changed